requirements.txt
notebooklm-mcp-server>=0.1.15
fastmcp
httpx
websocket-client
